+++ src/main/java/terra/backend/domain/cpu/controller/CpuUsageController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 import terra.backend.common.dto.response.ResponseDto;
 import terra.backend.domain.cpu.service.CpuService;
 import terra.backend.domain.cpu.validation.annotation.DateValidation;
-import terra.backend.domain.dto.response.CpuResponseDto;
+import terra.backend.domain.cpu.dto.response.CpuResponseDto;
 
 @Validated
 @RestController
