+++ src/main/java/terra/backend/domain/cpu/dto/response/CpuMinuteUsageResponse.java
@@ -1,4 +1,4 @@
-package terra.backend.domain.dto.response;
+package terra.backend.domain.cpu.dto.response;
 
 import java.time.format.DateTimeFormatter;
 import java.util.List;
