+++ src/main/java/terra/backend/domain/cpu/dto/response/CpuResponseDto.java
@@ -1,0 +1,3 @@
+package terra.backend.domain.cpu.dto.response;
+
+public interface CpuResponseDto {}
