+++ src/main/java/terra/backend/domain/cpu/service/CpuServiceImpl.java
@@ -1,7 +1,5 @@
 package terra.backend.domain.cpu.service;
 
-import static terra.backend.domain.dto.response.CpuHourUsageResponse.*;
-
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.temporal.Temporal;
@@ -12,6 +10,7 @@ import terra.backend.common.exception.exception.BusinessLogicException;
 import terra.backend.common.utils.DateUtils;
 import terra.backend.domain.cpu.cache.CpuCache;
 import terra.backend.domain.cpu.cache.dto.CpuUsage;
+import terra.backend.domain.cpu.dto.response.CpuHourUsageResponse.CpuHourUsageDto;
 import terra.backend.domain.cpu.entity.CpuDailyUsage;
 import terra.backend.domain.cpu.entity.CpuHourlyUsage;
 import terra.backend.domain.cpu.entity.CpuMinuteUsage;
@@ -21,12 +20,12 @@ import terra.backend.domain.cpu.repository.DailyUsageRepository;
 import terra.backend.domain.cpu.repository.HourlyUsageRepository;
 import terra.backend.domain.cpu.repository.MinuteUsageRepository;
 import terra.backend.domain.cpu.validation.enums.DateValidType;
-import terra.backend.domain.dto.response.CpuDailyUsageResponse;
-import terra.backend.domain.dto.response.CpuDailyUsageResponse.CpuDailyUsageDto;
-import terra.backend.domain.dto.response.CpuHourUsageResponse;
-import terra.backend.domain.dto.response.CpuMinuteUsageResponse;
-import terra.backend.domain.dto.response.CpuMinuteUsageResponse.CpuMinuteUsageDto;
-import terra.backend.domain.dto.response.CpuResponseDto;
+import terra.backend.domain.cpu.dto.response.CpuDailyUsageResponse;
+import terra.backend.domain.cpu.dto.response.CpuDailyUsageResponse.CpuDailyUsageDto;
+import terra.backend.domain.cpu.dto.response.CpuHourUsageResponse;
+import terra.backend.domain.cpu.dto.response.CpuMinuteUsageResponse;
+import terra.backend.domain.cpu.dto.response.CpuMinuteUsageResponse.CpuMinuteUsageDto;
+import terra.backend.domain.cpu.dto.response.CpuResponseDto;
 
 @Service
 @Transactional(readOnly = true)
