--- src/main/java/terra/backend/domain/dto/response/CpuResponseDto.java
@@ -1,3 +1,0 @@
-package terra.backend.domain.dto.response;
-
-public interface CpuResponseDto {}
